Dev/fix windows warning

+++ maptk/plugins/vxl/triangulate_landmarks.cxx
@@ -207,7 +207,7 @@ triangulate_landmarks
         auto lm = std::make_shared<vital::landmark_d>();
         lm->set_loc(vital::vector_3d(pt3d.x(), pt3d.y(), pt3d.z()));
         lm->set_covar(covariance_3d(error));
-        lm->set_observations(lm_cams.size());
+        lm->set_observations(static_cast<unsigned int>(lm_cams.size()));
         triangulated_lms[p.first] = lm;
       }
     }
